news rendering based on users choices
