+++ client/components/index.js
@@ -8,6 +8,5 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Signup} from './signup-form'
 export {Login} from './login-form'
-export {default as AllArticles} from './allArticles'
 export {default as ChoiceForm} from './choiceForm'
 export {default as UserArticles} from './userArticles'
