+++ client/components/navbar.js
@@ -13,7 +13,6 @@ const Navbar = ({ handleClick, isLoggedIn }) => (
           {/* The navbar will show these links after you log in */}
           <Link to="/home">Home</Link>
           <Link to="/account-settings">My Account</Link>
-          <Link to="/top-news">Top Headlines</Link>
           <Link to="#" onClick={handleClick}>Logout</Link>
         </div>
       }
