+++ client/components/user-home.js
@@ -2,7 +2,6 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {UserArticles} from '../components'
-import { UserNav } from './userNav';
 
 /**
  * COMPONENT
