+++ client/store/index.js
@@ -3,8 +3,7 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import { newsReducer } from "./allArticles";
-import { articlesByCatReducer, articlesByKeyReducer } from './userArticles';
+import { articlesByCatReducer, articlesByKeyReducer, newsReducer } from './userArticles';
 
 const reducer = combineReducers({user, articles: newsReducer, articlesByCategory: articlesByCatReducer, articlesByKeyword: articlesByKeyReducer})
 const middleware = composeWithDevTools(applyMiddleware(
